Refactored SeriesDatum, SeriesDatumConfig, SelectionModel, and initial selection behavior.
PiperOrigin-RevId: 214069871
