+++ charts_common/lib/src/chart/bar/bar_renderer.dart
@@ -32,8 +32,8 @@ import '../cartesian/axis/axis.dart'
 import '../common/canvas_shapes.dart' show CanvasBarStack, CanvasRect;
 import '../common/chart_canvas.dart' show ChartCanvas, FillPatternType;
 import '../common/datum_details.dart' show DatumDetails;
-import '../common/processed_series.dart'
-    show ImmutableSeries, MutableSeries, SeriesDatum;
+import '../common/processed_series.dart' show ImmutableSeries, MutableSeries;
+import '../common/series_datum.dart' show SeriesDatum;
 import '../../common/color.dart' show Color;
 
 /// Renders series data as a series of bars.
