+++ charts_common/test/chart/cartesian/cartesian_renderer_test.dart
@@ -20,6 +20,7 @@ import 'package:charts_common/src/chart/cartesian/cartesian_renderer.dart';
 import 'package:charts_common/src/chart/common/chart_canvas.dart';
 import 'package:charts_common/src/chart/common/datum_details.dart';
 import 'package:charts_common/src/chart/common/processed_series.dart';
+import 'package:charts_common/src/chart/common/series_datum.dart';
 import 'package:charts_common/src/common/symbol_renderer.dart';
 
 import 'package:mockito/mockito.dart';
