+++ charts_common/test/chart/common/behavior/selection/lock_selection_test.dart
@@ -39,7 +39,7 @@ class MockChart extends Mock implements BaseChart {
   }
 }
 
-class MockSelectionModel extends Mock implements SelectionModel {
+class MockSelectionModel extends Mock implements MutableSelectionModel {
   bool locked = false;
 }
