+++ charts_common/test/chart/common/behavior/selection/select_nearest_test.dart
@@ -20,6 +20,7 @@ import 'package:charts_common/src/chart/common/behavior/selection/select_nearest
 import 'package:charts_common/src/chart/common/behavior/selection/selection_trigger.dart';
 import 'package:charts_common/src/chart/common/datum_details.dart';
 import 'package:charts_common/src/chart/common/processed_series.dart';
+import 'package:charts_common/src/chart/common/series_datum.dart';
 import 'package:charts_common/src/chart/common/selection_model/selection_model.dart';
 import 'package:charts_common/src/common/gesture_listener.dart';
 import 'package:charts_common/src/data/series.dart';
@@ -43,7 +44,8 @@ class MockChart extends Mock implements BaseChart<String> {
   }
 }
 
-class MockSelectionModel extends Mock implements SelectionModel<String> {}
+class MockSelectionModel extends Mock implements MutableSelectionModel<String> {
+}
 
 void main() {
   MockChart _chart;
