+++ charts_common/test/chart/common/behavior/series_legend_behavior_test.dart
@@ -16,6 +16,7 @@
 import 'package:charts_common/src/chart/cartesian/axis/axis.dart';
 import 'package:charts_common/src/chart/common/base_chart.dart';
 import 'package:charts_common/src/chart/common/processed_series.dart';
+import 'package:charts_common/src/chart/common/series_datum.dart';
 import 'package:charts_common/src/chart/common/series_renderer.dart';
 import 'package:charts_common/src/chart/common/behavior/legend/legend.dart';
 import 'package:charts_common/src/chart/common/behavior/legend/legend_entry_generator.dart';
